[RESTEASY-1651] Enhance UndertowTest and HeadersInViewResponseTest
