+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/core/servlet/UndertowTest.java
@@ -43,7 +43,7 @@ public class UndertowTest {
                 new PropertyPermission("ipv6", "read"),
                 new RuntimePermission("getenv.RESTEASY_PORT"),
                 new PropertyPermission("org.jboss.resteasy.port", "read"),
-                new SocketPermission(PortProviderUtil.getHost(), "connect,resolve")), "permissions.xml");
+                new SocketPermission("[" + PortProviderUtil.getHost() + "]", "connect,resolve")), "permissions.xml");
         return war;
     }
