+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/providers/file/TempFileDeletionTest.java
@@ -11,6 +11,7 @@ import javax.ws.rs.core.Response;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.resteasy.category.NotForForwardCompatibility;
 import org.jboss.resteasy.test.providers.file.resource.TempFileDeletionResource;
 import org.jboss.resteasy.util.HttpResponseCodes;
 import org.jboss.resteasy.utils.PortProviderUtil;
@@ -20,6 +21,7 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.junit.experimental.categories.Category;
 
 /**
  * @tpSubChapter File provider
@@ -37,9 +39,13 @@ public class TempFileDeletionTest {
     }
 
     /**
+     * @tpTestDetails Resource method contains parameter of the type File. This triggers File provider, which creates
+     * temporary file on the server side, which is automatically deleted in the end of the resource method invocation.
+     * @tpInfo Regression test for RESTEASY-1464
      * @tpSince RESTEasy 3.1.3.Final
      */
     @Test
+    @Category(NotForForwardCompatibility.class)
     public void testDeleteOnServer() throws Exception {
         Client client = ClientBuilder.newClient();
         WebTarget base = client.target(PortProviderUtil.generateURL("/test/post", TempFileDeletionTest.class.getSimpleName()));
